Fix truncation on initialization for math constants Short some ints

+++ System/Constants.h
@@ -4,10 +4,10 @@
 namespace RTE {
 
 #pragma region Graphics Constants
-	static constexpr unsigned int c_PaletteEntriesNumber = 256; //!< Number of indexes in the graphics palette.
-	static constexpr unsigned int c_MOIDLayerBitDepth = 16; //!< Bit depth of MOID layer bitmap.
+	static constexpr unsigned short c_PaletteEntriesNumber = 256; //!< Number of indexes in the graphics palette.
+	static constexpr unsigned short c_MOIDLayerBitDepth = 16; //!< Bit depth of MOID layer bitmap.
 
-	static constexpr unsigned int c_GoldMaterialID = 2; //!< Index of gold material in the material palette.
+	static constexpr unsigned short c_GoldMaterialID = 2; //!< Index of gold material in the material palette.
 
 	// GUI colors
 	#define c_GUIColorWhite makecol(255, 255, 255)
@@ -25,12 +25,12 @@ namespace RTE {
 #pragma endregion
 
 #pragma region Math Constants
-	static constexpr float c_TwoPI = 6.2831853;
-	static constexpr float c_PI = 3.1415926;
-	static constexpr float c_HalfPI = 1.5707963;
-	static constexpr float c_QuarterPI = 0.7853982;
-	static constexpr float c_EighthPI = 0.3926991;
-	static constexpr float c_SixteenthPI = 0.1963495;
+	static constexpr float c_TwoPI = 6.2831853F;
+	static constexpr float c_PI = 3.1415926F;
+	static constexpr float c_HalfPI = 1.5707963F;
+	static constexpr float c_QuarterPI = 0.7853982F;
+	static constexpr float c_EighthPI = 0.3926991F;
+	static constexpr float c_SixteenthPI = 0.1963495F;
 #pragma endregion
 
 #pragma region Network Constants
